reference_code/T2/requirements.txt
# Core Dependencies
selenium>=4.0.0,<5.0.0
python-dotenv>=0.21.0,<2.0.0
colorama>=0.4.4,<1.0.0
pydantic>=2.0.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
typing-extensions>=4.0.0,<5.0.0

# Development Dependencies
pytest>=7.0.0,<8.0.0
pytest-cov>=3.0.0,<4.0.0
black>=22.0.0,<23.0.0
flake8>=5.0.0,<6.0.0
mypy>=1.0.0,<2.0.0
pytest-mock>=3.0.0,<4.0.0
